requirements.txt
boto3==1.34.62
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pyparsing==3.1.2
python-docx==1.1.0
typing-extensions==4.10.0
PyYAML==6.0.1
SQLAlchemy==2.0.28
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-core==0.1.33
langsmith==0.1.31
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.14.2
pydantic==2.6.4
pydantic-core==2.16.3
requests==2.31.0
sniffio==1.3.1
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
yarl==1.9.4
python-dotenv==1.0.1